Add files via upload
The provided MergeH.ino file contains a NodeMCU code.
The code incorporates a multiplexing algorithm which is used to take 3 analogue inputs from node mcu. the multiplexer used is 4051 ic wherein the selection lines are connected to D0 D1 D2 pins of the node.

This Arduino code reads in gas sensor levels of Ozone, diHydrogen Sulphide and Carbon Monoxide and updated the real-time readings on the firebase.